+++ site/dat/docs/Changelog.md
@@ -4,6 +4,7 @@
  - fix hamcrest installation for Java-based Selenium tests
  - fix Jmeter fail with unusual jmx property type
  - add variable sending example into Gatling documentation
+ - fix results time and add widget for ApacheBenchmark
 
 ## 1.6.7  <sup>23 aug 2016</sup>
  - add `cloud` and `local` aliases